Removes boost converter placeholder since it won't be possible to use it
The stamp can only provide 500mA@3.3v, which gives us 1.65W. Assuming a
85% efficiency of the boost regulator, that would give us 1.4W on the
output, or 280mA@5V. We need 320mA to run the 64 WS2812C, so it doesn't
seem to be possible to run these from a LiPO battery unless we rely on
the software never running all the LEDs at full brightness, but that
doesn't seem like a reliable option.

We'll still use the WS2812C instead of the B for the lower power
consumption, though. Since we don't need a lot of brightness, running
the matrix with 320mA should make it possible for the whole thing to run
from a standard USB connection.