Adjusts LED matrix placement
Because of the NES connector and it's through hole leads it's really
hard to keep it symmetric while keeping it under 100mm x 100mm*. The
solution I found is to keep the spacing between the LEDs and the edges
the same in all but the side where the connectors are. This would
probably be better in the bottom, but then the cables would have nowhere
to exit from.

When designing a 3D printed enclosure we can try to hide this asymmetry.

* Since this is a 4 layer board, JLCPC's engineering fee will spike
  after this size threshold.